v0.36.4-beta

+++ package/my/luci-app-openclash/files/usr/share/openclash/yml_proxys_set.sh
@@ -1,16 +1,109 @@
-#!/bin/sh /etc/rc.common
+#!/bin/sh
+. /lib/functions.sh
+
 status=$(ps|grep -c /usr/share/openclash/yml_proxys_set.sh)
 [ "$status" -gt "3" ] && exit 0
 
 START_LOG="/tmp/openclash_start.log"
 SERVER_FILE="/tmp/yaml_servers.yaml"
+PROXY_PROVIDER_FILE="/tmp/yaml_provider.yaml"
+servers_if_update=$(uci get openclash.config.servers_if_update 2>/dev/null)
+config_auto_update=$(uci get openclash.config.auto_update 2>/dev/null)
+CONFIG_FILE=$(uci get openclash.config.config_path 2>/dev/null)
+CONFIG_NAME=$(echo $CONFIG_FILE |awk -F '/' '{print $5}' 2>/dev/null)
+UPDATE_CONFIG_FILE=$(uci get openclash.config.config_update_path 2>/dev/null)
+UPDATE_CONFIG_NAME=$(echo $UPDATE_CONFIG_FILE |awk -F '/' '{print $5}' 2>/dev/null)
 
+if [ ! -z "$UPDATE_CONFIG_FILE" ]; then
+   CONFIG_FILE="$UPDATE_CONFIG_FILE"
+   CONFIG_NAME="$UPDATE_CONFIG_NAME"
+fi
+
+if [ -z "$CONFIG_FILE" ]; then
+	CONFIG_FILE="/etc/openclash/config/$(ls -lt /etc/openclash/config/ | grep -E '.yaml|.yml' | head -n 1 |awk '{print $9}')"
+	CONFIG_NAME=$(echo $CONFIG_FILE |awk -F '/' '{print $5}' 2>/dev/null)
+fi
+
+if [ -z "$CONFIG_NAME" ]; then
+   CONFIG_FILE="/etc/openclash/config/config.yaml"
+   CONFIG_NAME="config.yaml"
+fi
+
+#写入代理集到配置文件
+yml_proxy_provider_set()
+{
+   local section="$1"
+   config_get_bool "enabled" "$section" "enabled" "1"
+   config_get "config" "$section" "config" ""
+   config_get "type" "$section" "type" ""
+   config_get "name" "$section" "name" ""
+   config_get "path" "$section" "path" ""
+   config_get "provider_url" "$section" "provider_url" ""
+   config_get "provider_interval" "$section" "provider_interval" ""
+   config_get "health_check" "$section" "health_check" ""
+   config_get "health_check_url" "$section" "health_check_url" ""
+   config_get "health_check_interval" "$section" "health_check_interval" ""
+   
+   if [ ! -z "$if_game_proxy" ] && [ "$if_game_proxy" != "$name" ] && [ "$if_game_proxy_type" = "proxy-provider" ]; then
+      return
+   fi
+   
+   if [ ! -z "$config" ] && [ "$config" != "$CONFIG_NAME" ] && [ "$config" != "all" ]; then
+      return
+   fi
+   
+   if [ "$enabled" = "0" ]; then
+      return
+   fi
+
+   if [ -z "$type" ]; then
+      return
+   fi
+   
+   if [ -z "$name" ]; then
+      return
+   fi
+   
+   if [ "$path" != "./proxy_provider/$name.yaml" ] && [ "$type" = "http" ]; then
+      path="./proxy_provider/$name.yaml"
+   elif [ -z "$path" ]; then
+      return
+   fi
+   
+   if [ -z "$health_check" ]; then
+      return
+   fi
+   
+   echo "正在写入【$type】-【$name】代理集到配置文件【$CONFIG_NAME】..." >$START_LOG
+   echo "$name" >> /tmp/Proxy_Provider
+   
+cat >> "$PROXY_PROVIDER_FILE" <<-EOF
+  $name:
+    type: $type
+    path: $path
+EOF
+   if [ ! -z "$provider_url" ]; then
+cat >> "$PROXY_PROVIDER_FILE" <<-EOF
+    url: $provider_url
+    interval: $provider_interval
+EOF
+   fi
+cat >> "$PROXY_PROVIDER_FILE" <<-EOF
+    health-check:
+      enable: $health_check
+      url: $health_check_url
+      interval: $health_check_interval
+EOF
+
+}
 
 #写入服务器节点到配置文件
 yml_servers_set()
 {
 
    local section="$1"
+   config_get_bool "enabled" "$section" "enabled" "1"
+   config_get "config" "$section" "config" ""
    config_get "type" "$section" "type" ""
    config_get "name" "$section" "name" ""
    config_get "server" "$section" "server" ""
@@ -22,6 +115,7 @@ yml_servers_set()
    config_get "obfs" "$section" "obfs" ""
    config_get "obfs_vmess" "$section" "obfs_vmess" ""
    config_get "host" "$section" "host" ""
+   config_get "mux" "$section" "mux" ""
    config_get "custom" "$section" "custom" ""
    config_get "tls" "$section" "tls" ""
    config_get "skip_cert_verify" "$section" "skip_cert_verify" ""
@@ -30,9 +124,26 @@ yml_servers_set()
    config_get "uuid" "$section" "uuid" ""
    config_get "auth_name" "$section" "auth_name" ""
    config_get "auth_pass" "$section" "auth_pass" ""
-
+   config_get "psk" "$section" "psk" ""
+   config_get "obfs_snell" "$section" "obfs_snell" ""
+   
+   if [ ! -z "$if_game_proxy" ] && [ "$if_game_proxy" != "$name" ] && [ "$if_game_proxy_type" = "proxy" ]; then
+      return
+   fi
+   
+   if [ ! -z "$config" ] && [ "$config" != "$CONFIG_NAME" ] && [ "$config" != "all" ]; then
+      return
+   fi
+   
+   if [ "$enabled" = "0" ]; then
+      return
+   fi
 
    if [ -z "$type" ]; then
+      return
+   fi
+   
+   if [ -z "$name" ]; then
       return
    fi
    
@@ -48,6 +159,8 @@ yml_servers_set()
       return
    fi
    
+   echo "正在写入【$type】-【$name】节点到配置文件【$CONFIG_NAME】..." >$START_LOG
+   
    if [ "$obfs" != "none" ]; then
       if [ "$obfs" = "websocket" ]; then
          obfss="plugin: v2ray-plugin"
@@ -60,8 +173,14 @@ yml_servers_set()
       udp=", udp: $udp"
    fi
    
+   if [ "$obfs_snell" = "none" ]; then
+      obfs_snell=""
+   fi
+   
    if [ "$obfs_vmess" != "none" ]; then
-         obfs_vmess=", network: ws"
+      obfs_vmess=", network: ws"
+   else
+      obfs_vmess=""
    fi
    
    if [ ! -z "$host" ]; then
@@ -93,7 +212,7 @@ yml_servers_set()
    fi
 
    if [ "$type" = "ss" ] && [ "$obfs" = "none" ]; then
-      echo "- { name: \"$name\", type: $type, server: $server, port: $port, cipher: $cipher, password: "$password"$udp }" >>$SERVER_FILE
+      echo "- { name: \"$name\", type: $type, server: $server, port: $port, cipher: $cipher, password: \"$password\"$udp }" >>$SERVER_FILE
    elif [ "$type" = "ss" ] && [ "$obfs" != "none" ]; then
 cat >> "$SERVER_FILE" <<-EOF
 - name: "$name"
@@ -131,6 +250,11 @@ cat >> "$SERVER_FILE" <<-EOF
     $path
 EOF
   fi
+  if [ ! -z "$mux" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+    mux: $mux
+EOF
+  fi
   if [ ! -z "$custom" ]; then
 cat >> "$SERVER_FILE" <<-EOF
     headers:
@@ -146,30 +270,76 @@ EOF
    if [ "$type" = "socks5" ] || [ "$type" = "http" ]; then
       echo "- { name: \"$name\", type: $type, server: $server, port: $port, username: $auth_name, password: $auth_pass$skip_cert_verify$tls }" >>$SERVER_FILE
    fi
+   
+   if [ "$type" = "snell" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+- name: "$name"
+  type: $type
+  server: $server
+  port: $port
+  psk: $psk
+EOF
+   if [ "$obfs_snell" != "none" ] && [ ! -z "$host" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  obfs-opts:
+    mode: $obfs_snell
+    $host
+EOF
+   fi
+   fi
 
 }
 
 
 #创建配置文件
-echo "开始更新配置文件节点..." >$START_LOG
-echo "Proxy:" >$SERVER_FILE
+if_game_proxy="$1"
+if_game_proxy_type="$2"
+#proxy-provider
+echo "开始写入配置文件【$CONFIG_NAME】的代理集信息..." >$START_LOG
+echo "proxy-provider:" >$PROXY_PROVIDER_FILE
+rm -rf /tmp/Proxy_Provider
 config_load "openclash"
-config_foreach yml_servers_set "servers"
-echo "Proxy Group:" >>$SERVER_FILE
+config_foreach yml_proxy_provider_set "proxy-provider"
+sed -i "s/^ \{0,\}/  - /" /tmp/Proxy_Provider 2>/dev/null #添加参数
+if [ "$(grep "-" /tmp/Proxy_Provider 2>/dev/null |wc -l)" -eq 0 ]; then
+   rm -rf $PROXY_PROVIDER_FILE
+   rm -rf /tmp/Proxy_Provider
+fi
+
+#proxy
 rule_sources=$(uci get openclash.config.rule_sources 2>/dev/null)
 create_config=$(uci get openclash.config.create_config 2>/dev/null)
-egrep '^ {0,}-' $SERVER_FILE |grep name: |awk -F 'name: ' '{print $2}' |sed 's/,.*//' >/tmp/Proxy_Server 2>&1
-sed -i "s/^ \{0,\}/  - /" /tmp/Proxy_Server 2>/dev/null #添加参数
+echo "开始写入配置文件【$CONFIG_NAME】的服务器节点信息..." >$START_LOG
+echo "Proxy:" >$SERVER_FILE
+config_foreach yml_servers_set "servers"
+egrep '^ {0,}-' $SERVER_FILE |grep name: |awk -F 'name: ' '{print $2}' |sed 's/,.*//' 2>/dev/null >/tmp/Proxy_Server 2>&1
+if [ -s "/tmp/Proxy_Server" ]; then
+   sed -i "s/^ \{0,\}/  - /" /tmp/Proxy_Server 2>/dev/null #添加参数
+else
+   rm -rf $SERVER_FILE
+   rm -rf /tmp/Proxy_Server
+fi
 
-
-if [ "$rule_sources" = "ConnersHua" ]; then
+#一键创建配置文件
+if [ "$rule_sources" = "ConnersHua" ] && [ "$servers_if_update" != "1" ] && [ -z "$if_game_proxy" ]; then
 echo "使用ConnersHua规则创建中..." >$START_LOG
+echo "Proxy Group:" >>$SERVER_FILE
 cat >> "$SERVER_FILE" <<-EOF
 - name: Auto - UrlTest
   type: url-test
+EOF
+if [ -f "/tmp/Proxy_Server" ]; then
+cat >> "$SERVER_FILE" <<-EOF
   proxies:
 EOF
+fi
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
   url: http://www.gstatic.com/generate_204
   interval: "600"
@@ -180,6 +350,12 @@ cat >> "$SERVER_FILE" <<-EOF
   - DIRECT
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
 - name: Domestic
   type: select
@@ -210,6 +386,12 @@ cat >> "$SERVER_FILE" <<-EOF
   - Proxy
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
 - name: GlobalTV
   type: select
@@ -218,6 +400,12 @@ cat >> "$SERVER_FILE" <<-EOF
   - DIRECT
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 uci set openclash.config.rule_source="ConnersHua"
 uci set openclash.config.GlobalTV="GlobalTV"
 uci set openclash.config.AsianTV="AsianTV"
@@ -226,14 +414,33 @@ uci set openclash.config.Apple="Apple"
 uci set openclash.config.AdBlock="AdBlock"
 uci set openclash.config.Domestic="Domestic"
 uci set openclash.config.Others="Others"
-elif [ "$rule_sources" = "lhie1" ]; then
+[ "$config_auto_update" -eq 1 ] && {
+	uci set openclash.config.servers_update="1"
+	uci del openclash.config.new_servers_group >/dev/null 2>&1
+	uci add_list openclash.config.new_servers_group="Auto - UrlTest"
+	uci add_list openclash.config.new_servers_group="Proxy"
+ 	uci add_list openclash.config.new_servers_group="AsianTV"
+	uci add_list openclash.config.new_servers_group="GlobalTV"
+}
+elif [ "$rule_sources" = "lhie1" ] && [ "$servers_if_update" != "1" ] && [ -z "$if_game_proxy" ]; then
 echo "使用lhie1规则创建中..." >$START_LOG
+echo "Proxy Group:" >>$SERVER_FILE
 cat >> "$SERVER_FILE" <<-EOF
 - name: Auto - UrlTest
   type: url-test
+EOF
+if [ -f "/tmp/Proxy_Server" ]; then
+cat >> "$SERVER_FILE" <<-EOF
   proxies:
 EOF
+fi
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
   url: http://www.gstatic.com/generate_204
   interval: "600"
@@ -244,6 +451,12 @@ cat >> "$SERVER_FILE" <<-EOF
   - DIRECT
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
 - name: Domestic
   type: select
@@ -256,6 +469,17 @@ cat >> "$SERVER_FILE" <<-EOF
   - Proxy
   - DIRECT
   - Domestic
+- name: Apple
+  type: select
+  proxies:
+  - DIRECT
+  - Proxy
+- name: AdBlock
+  type: select
+  proxies:
+  - REJECT
+  - DIRECT
+  - Proxy
 - name: AsianTV
   type: select
   proxies:
@@ -263,6 +487,12 @@ cat >> "$SERVER_FILE" <<-EOF
   - Proxy
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
 - name: GlobalTV
   type: select
@@ -271,20 +501,95 @@ cat >> "$SERVER_FILE" <<-EOF
   - DIRECT
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
+cat >> "$SERVER_FILE" <<-EOF
+- name: Speedtest
+  type: select
+  proxies:
+  - Proxy
+  - DIRECT
+EOF
+cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
+cat >> "$SERVER_FILE" <<-EOF
+- name: Telegram
+  type: select
+  proxies:
+  - Proxy
+  - DIRECT
+EOF
+cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
+cat >> "$SERVER_FILE" <<-EOF
+- name: Netease Music
+  type: select
+  proxies:
+  - DIRECT
+  - Proxy
+EOF
+cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 uci set openclash.config.rule_source="lhie1"
 uci set openclash.config.GlobalTV="GlobalTV"
 uci set openclash.config.AsianTV="AsianTV"
 uci set openclash.config.Proxy="Proxy"
+uci set openclash.config.Apple="Apple"
+uci set openclash.config.AdBlock="AdBlock"
+uci set openclash.config.Netease_Music="Netease Music"
+uci set openclash.config.Speedtest="Speedtest"
+uci set openclash.config.Telegram="Telegram"
 uci set openclash.config.Domestic="Domestic"
 uci set openclash.config.Others="Others"
-elif [ "$rule_sources" = "ConnersHua_return" ]; then
+[ "$config_auto_update" -eq 1 ] && {
+	uci set openclash.config.servers_update="1"
+	uci del openclash.config.new_servers_group >/dev/null 2>&1
+	uci add_list openclash.config.new_servers_group="Auto - UrlTest"
+	uci add_list openclash.config.new_servers_group="Proxy"
+ 	uci add_list openclash.config.new_servers_group="AsianTV"
+	uci add_list openclash.config.new_servers_group="GlobalTV"
+	uci add_list openclash.config.new_servers_group="Telegram"
+	uci add_list openclash.config.new_servers_group="Speedtest"
+	uci add_list openclash.config.new_servers_group="Netease Music"
+}
+elif [ "$rule_sources" = "ConnersHua_return" ] && [ "$servers_if_update" != "1" ] && [ -z "$if_game_proxy" ]; then
 echo "使用ConnersHua回国规则创建中..." >$START_LOG
+echo "Proxy Group:" >>$SERVER_FILE
 cat >> "$SERVER_FILE" <<-EOF
 - name: Auto - UrlTest
   type: url-test
+EOF
+if [ -f "/tmp/Proxy_Server" ]; then
+cat >> "$SERVER_FILE" <<-EOF
   proxies:
 EOF
+fi
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
   url: http://www.gstatic.com/generate_204
   interval: "600"
@@ -295,6 +600,12 @@ cat >> "$SERVER_FILE" <<-EOF
   - DIRECT
 EOF
 cat /tmp/Proxy_Server >> $SERVER_FILE 2>/dev/null
+if [ -f "/tmp/Proxy_Provider" ]; then
+cat >> "$SERVER_FILE" <<-EOF
+  use:
+EOF
+fi
+cat /tmp/Proxy_Provider >> $SERVER_FILE 2>/dev/null
 cat >> "$SERVER_FILE" <<-EOF
 - name: Others
   type: select
@@ -305,21 +616,61 @@ EOF
 uci set openclash.config.rule_source="ConnersHua_return"
 uci set openclash.config.Proxy="Proxy"
 uci set openclash.config.Others="Others"
+[ "$config_auto_update" -eq 1 ] && {
+	uci set openclash.config.servers_update="1"
+	uci del openclash.config.new_servers_group >/dev/null 2>&1
+	uci add_list openclash.config.new_servers_group="Auto - UrlTest"
+	uci add_list openclash.config.new_servers_group="Proxy"
+}
 fi
-if [ "$create_config" != "0" ]; then
+
+if [ "$create_config" != "0" ] && [ "$servers_if_update" != "1" ] && [ -z "$if_game_proxy" ]; then
    echo "Rule:" >>$SERVER_FILE
-   uci commit openclash
-   cat "$SERVER_FILE" > "/etc/openclash/config.yaml" 2>/dev/null
-else
-   echo "正在更新配置文件服务器节点信息..." >$START_LOG
-   /usr/share/openclash/yml_groups_set.sh
-   sed -i '/^ \{0,\}Proxy:/i\#change server#' "/etc/openclash/config.yaml" 2>/dev/null
-   sed -i '/^ \{0,\}Proxy:/,/^ \{0,\}Rule:/d' "/etc/openclash/config.yaml" 2>/dev/null
-   sed -i '/#change server#/r/tmp/yaml_servers.yaml' "/etc/openclash/config.yaml" 2>/dev/null
-   sed -i '/Proxy Group:/r/tmp/yaml_groups.yaml' "/etc/openclash/config.yaml" 2>/dev/null
-   sed -i '/#change server#/d' "/etc/openclash/config.yaml" 2>/dev/null
+   echo "配置文件【$CONFIG_NAME】创建完成，正在更新服务器、代理集、策略组信息..." >$START_LOG
+   cat "$PROXY_PROVIDER_FILE" > "$CONFIG_FILE" 2>/dev/null
+   cat "$SERVER_FILE" >> "$CONFIG_FILE" 2>/dev/null
+   /usr/share/openclash/yml_groups_get.sh >/dev/null 2>&1
+elif [ -z "$if_game_proxy" ]; then
+   echo "服务器、代理集、策略组信息修改完成，正在更新配置文件【$CONFIG_NAME】..." >$START_LOG
+   #判断各个区位置
+   proxy_len=$(sed -n '/^Proxy:/=' "$CONFIG_FILE" 2>/dev/null)
+   group_len=$(sed -n '/^ \{0,\}Proxy Group:/=' "$CONFIG_FILE" 2>/dev/null)
+   provider_len=$(sed -n '/^proxy-provider:/=' "$CONFIG_FILE" 2>/dev/null)
+   if [ "$provider_len" -le "$proxy_len" ]; then
+      sed -i '/^ \{0,\}proxy-provider:/i\#change server#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}Rule:/i\#change server end#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}proxy-provider:/,/#change server end#/d' "$CONFIG_FILE" 2>/dev/null
+   elif [ "$provider_len" -le "$group_len" ] && [ -z "$proxy_len" ]; then
+      sed -i '/^ \{0,\}proxy-provider:/i\#change server#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}Rule:/i\#change server end#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}proxy-provider:/,/#change server end#/d' "$CONFIG_FILE" 2>/dev/null
+   elif [ "$provider_len" -ge "$group_len" ] && [ -z "$proxy_len" ]; then
+      sed -i '/^ \{0,\}Proxy Group:/i\#change server#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}Rule:/i\#change server end#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}Proxy Group:/,/#change server end#/d' "$CONFIG_FILE" 2>/dev/null
+   else
+      sed -i '/^ \{0,\}Proxy:/i\#change server#' "$CONFIG_FILE" 2>/dev/null
+   	  sed -i '/^ \{0,\}Rule:/i\#change server end#' "$CONFIG_FILE" 2>/dev/null
+      sed -i '/^ \{0,\}Proxy:/,/#change server end#/d' "$CONFIG_FILE" 2>/dev/null
+   fi
+
+   sed -i '/#change server#/r/tmp/yaml_groups.yaml' "$CONFIG_FILE" 2>/dev/null
+   sed -i '/#change server#/r/tmp/yaml_servers.yaml' "$CONFIG_FILE" 2>/dev/null
+   sed -i '/#change server#/r/tmp/yaml_provider.yaml' "$CONFIG_FILE" 2>/dev/null
+   sed -i '/#change server#/d' "$CONFIG_FILE" 2>/dev/null
 fi
-echo "配置文件更新完成！" >$START_LOG
-rm -rf $SERVER_FILE 2>/dev/null
+echo "配置文件【$CONFIG_NAME】写入完成！" >$START_LOG
+sleep 3
+echo "" >$START_LOG
+if [ -z "$if_game_proxy" ]; then
+   rm -rf $SERVER_FILE 2>/dev/null
+   rm -rf $PROXY_PROVIDER_FILE 2>/dev/null
+   rm -rf /tmp/yaml_groups.yaml 2>/dev/null
+fi
 rm -rf /tmp/Proxy_Server 2>/dev/null
-rm -rf /tmp/yaml_groups.yaml 2>/dev/null
+rm -rf /tmp/Proxy_Provider 2>/dev/null
+uci set openclash.config.enable=1 2>/dev/null
+[ "$(uci get openclash.config.servers_if_update)" == "0" ] && [ -z "$if_game_proxy" ] && /etc/init.d/openclash restart >/dev/null 2>&1
+uci set openclash.config.servers_if_update=0
+uci commit openclash
+
